Introduce interface for data to be uploaded
Summary:
Currently we pass around data to be uploaded as just a
ThrowingSupplier<InputStream, IOException>. This makes it difficult to
pass around any additional data/behavior with that. In particular, when
we get an error uploading some blob, we'd like to understand what that
blob is and where it came from.

This diff changes that Supplier to an explicit new interface that we can
make richer in the future.

Since the new interface is still a functional interface with the same
signature, we don't need to change any actual implementations yet.

Reviewed By: bobyangyf

fbshipit-source-id: 5c879ad53b