Support identical native libraries in exopackage install
Summary:
The exopackage install process builds an immutable map from file hash to
shared object file.  Building this throws if there are multiple native
libs with identical content.  Make it a multimap handle this case
better.  When the time comes to actually install the libs, we just
choose one to install.

Reviewed By: cjhopman

fbshipit-source-id: ea538bdc77