Introduce android.sdk_path_search_order config option to define the order of search for Android SDK location
Summary:
Some users may prefer to have a different order in which Android SDK location is determined. This change add a configuration option "android.sdk_path_search_order" which allows to define a custom order for search as well as using custom env variables.

The entries in this options are either the names of env variables or a special token "<CONFIG>" which indicates that the location should be read from `.buckconfig` option (in `android.sdk_path`).

fbshipit-source-id: aae7ab4b90