Fix source ABI generation crash when module-info.java doesn't exist
Summary:
If we target Java 10+, a lookup sometimes happens for the `module-info` class, which causes a `NullPointerException` if this class doesn't exist (even though it's perfectly valid to not have one), as `ForwardingJavaFileObject` in the JDK doesn't support a null inner object.

Adding a null check in Buck code. I haven't added the same check to the other methods in `ListenableFileManager`, as I'd like to understand the scenarios if crashes do occur in the other cases.

Reviewed By: styurin

fbshipit-source-id: 1d107a9683