Restrict the # of pending uploads
Summary:
For RE, for each rule we construct a merkle tree of inputs and the
associated data (file inputs, serialized rule, some internal stuff) and
then upload all that data. For the actual execute request we only need a
single digest (and we need the list of outputs to handle the result).

While the upload is pending, each rule may hold a reference to some
large amount of data (1MB+). Since actions are computed very quickly and
build graphs may be very flat, we can end up with 1000s of rules in this
state fairly easily. That could require a ridiculous amount of memory.

This adds another configuration value
remoteexecution.concurrent_pending_uploads that limits how many rules
can be in this state. This is used for a JobLimiter that limits around
the beginning of action computation to the end of uploading (because
this large data is a result of action computation).

Reviewed By: alisdair04

fbshipit-source-id: bf99c29914