On failed actions, log the worker
Summary:
For actions that have an error we print the error message from the
server. For failed actions, we print just the stderr from the action
itself. Sometimes, failed actions are actually caused by a misbehaving
worker. To help identify that case, log the worker on failed actions.

I think we should improve this in the future to include it as part of
the printed message, but that's harder to do in a principled way. This
was easy.

Reviewed By: bobyangyf

fbshipit-source-id: 14676bdee9