Allow the cacheability of strip rules to be configured
Summary:
If we allow CxxLink to be uncached, we should also do the same for CxxStrip,
since strip usually runs on the result of a link.

Making this rule uncached mitigates problems involving the linker generating
non-reproducible results. Rules operating on outputs from one run of the link
rule (namely, the linkmap) and the cached result of strip from a different run
of the link rule may produce frankenbuilds or fail.

Reviewed By: cjhopman

fbshipit-source-id: 86aac257fe