Handle directories properly in BuckTargetLocator.
Summary:
Allow BuckTargetLocator to return a target pattern for directories,
and not just files.  In doing so, found/fixed issues related to
resolving patterns when a directory has the same filename as the
buck cell's `buildfile.name`.   (This happens in the buck repo on
case-insensitive filesystems, since the buildfile name is `BUCK`
and there exist several directories like `src/com/facebook/buck`.)

Reviewed By: IanChilds

fbshipit-source-id: 30c8ca259a