Reject rules that cannot be serialized
Summary:
Rather than needing to support all possible configurations of
ModernBuildRules, add a mechanism to inspect the rule early on and
reject it if necessary (that'll allow it to fallback to the same path as
non-MBR rules). We use this mechanism right now just to reject anything
that doesn't serialize successfully which handles our immediate needs,
but we could imagine using this in other ways as well.

Reviewed By: ruibm

fbshipit-source-id: 46dd871af7