Create symlinks of resources in sh_binary rule.
Summary:
sh_binary rule creates symlinks of each resource in a new directory under the target's gen path.  The absolute path of the symlink directory structure, as resolved at runtime, is exported as *$BUCK_RUNTIME_RESOURCES* in the sh script's environment.

The structure of the symlink paths follow the conventions below:
| Resource Type | Linked Path |
| -- | -- |
|Build target resources (same cell) |$BUCK_RUNTIME_RESOURCES/__default__/*{target}*|
|Build target resources (different cell)|$BUCK_RUNTIME_RESOURCES/__external__/*{cell}*/*{target}*|
|Path resources (sh_binary rule)|$BUCK_RUNTIME_RESOURCES/__default__/*{relative resource path}*|

If sh_binary detects that conflicting path links would be generated, it fails the build.

Reviewed By: philipjameson

fbshipit-source-id: 05c52c98db