Fix GWT annotation processing when using buck cache, by putting the dependency on the right location
Summary:
In previous commit, I've added the respected javaLibrary rule target dependency on the wrong rule.

Only the `GwtBinary` rule requires the java library, since only at this final stage the generated sources are used.

When the dependency was added directly to that library matching `GwtModule` then it failed to build when buck caching was enabled. Because when recompiling, it did not extracted the annotation generated sources to our `buck-out`

Reviewed By: styurin

fbshipit-source-id: f3260a1368