In GwtModule dependency traversing, make sure to depend on the real dependencies, and not the source abis.
Summary:
When traversing the GwtBinary module dependencies, make sure to traverse the real rule targets, and not their source abi flavors.

Because we need to collect all their sources.
This is similar to how native libraries traversing works for `JavaBinaryDescription`.

Without this fix, when enabled source abi compilation mode, I would get failures such as:
```
BUILT 7/8 JOBS 0.3s //:a
Loading inherited module 'com/example/gwt/A'
   Loading inherited module 'com.google.gwt.core.Core'
      [ERROR] Unable to find 'com/google/gwt/core/Core.gwt.xml' on your classpath; could be a typo, or maybe you forgot to include a classpath entry for source?
```
Since it didn't pushed the `:gwt-user` dependency to the `gwt_binary()` target, with the fix it works, since it is copied from `:a` target

Reviewed By: cjhopman

fbshipit-source-id: d4922508cd