Calculate TargetNode raw input hashes only when they are needed
Summary:
`buck targets` command relies on `TargetGraphHashing` to show the hashes of the targets. These hashes are eagerly calculated when `TargetNode`s are created in parser and the are not used in other places.

This change removes `rawInputsHashCode` field from `TargetNode` and changes `TargetGraphHashing` to calculate hashes by getting raw attributes of `TargetNode`s from the parser and performing hashing of those attributes.

This also solves the problem with `TargetNode`s that have configured attributes since the factory used to use `hashCode` of the source `RawTargetNode`.

Reviewed By: sbalabanov

fbshipit-source-id: 77dad47f7b