Be a bit more defensive about handling cancelled buildrules
Summary:
Changes the build engine to only accept failures from the
CachingBuildRuleBuilder that are of type BuildRuleFailedException and
thus enforces that the exception has the appropriate context.

Changes CachingBuildEngine to cancel CachingBuildRuleBuilder immediately
after adding it to the list of pending builders if there's currently a
failure.

Fixes a place in CachingBuildRuleBuilder where we weren't attaching
context to the failure.

Reviewed By: sbalabanov

fbshipit-source-id: 22331343f0