Fix handling of top level blocks in source-only ABI generation for Java 9+
Summary:
The behavior of `JavacTrees::getElement` in the Java Compiler Tree API has changed in newer versions of Java. In Java 8 and earlier, `getElement` would immediately return null when given a top level block (static initializer or anonymous block) as input. In Java 9+, `getElement` attempts to attribute the parent class, effectively starting an analyze phase (whereas we want to stop after parse and enter), which leads us to crash for various reasons.

Top level blocks are only used at runtime, and thus are irrelevant for source ABI generation. They should be skipped entirely. Add pre-filtering to fix this on Java 9+.

Reviewed By: jkeljo

fbshipit-source-id: 264d374911