Bring back our fork of ApkBuilder (without exposing APK compression level)
Summary:
Buck does APK building in-proc by calling into AOSP's `ApkBuilder`. `ApkBuilder` has a dependency on certain JCL classes which do not exist in Java 9+, which is problematic at runtime when building with a Java 10/11 build of Buck.

This change brings back our fork of this code (:(), which includes my fixes for these dependencies to ensure Java 9+ compatibility.

Reviewed By: cjhopman

fbshipit-source-id: 28dfd4ba1e