Fix importing Obj-C modular libraries into Swift
Summary:
There are two ways to import Objective-C code into Swift: briding headers, and modules. Bridging headers are worse in multiple ways:

- They're not "Swifty": there's no per-file `import`, the symbols are just magically available in every source file.
- Libraries built with them are not cacheable, because they include absolute paths. See [`SwiftCompile.java`](https://github.com/facebook/buck/blob/master/src/com/facebook/buck/swift/SwiftCompile.java#L301).

Therefore, it's better to use modules. In Buck, this is done by specifying `modular = True` in an `apple_library`, which will then generate a `.modulemap`, which Swift can understand and use to import the library. This works perfectly with `buck build`, but it doesn't currently work with `buck project`: only Swift libraries can be imported. The core reason for this is that the `SWIFT_INCLUDE_PATHS` for a library containing Swift is currently only set to `$BUILT_PRODUCTS_DIR`. There are a few changes needed to fix this:

- When we decide whether or not to create a header symlink tree, we should do it for all modular `apple_library` targets. The method that is called (`createHeaderSymlinkTree`) is also responsible for creating the `.modulemap` file that we need.
- Similar to building the list of header search paths for Objective-C, we now build a list of Swift import paths. This is then set in the `SWIFT_INCLUDE_PATHS` Xcode build setting.

After these changes, I can now `import` Objective-C code into a Swift library.

Reviewed By: williamtwilson

fbshipit-source-id: e3cfade749