Adds support for notification payload files in xcode_workspace_config
Summary: This adds the ability to add notification_payload_file to an xcode_workspace_config and have it show up as selected in the scheme. This is helpful when using xcode_workspace_config with extra_schemes to define notification schemes and have them appear in Xcode after running buck project.

Reviewed By: skotchvail

fbshipit-source-id: 4ae45f1cec