Calculate TargetNode raw input hashes only when they are needed
Summary:
`buck targets` command relies on `TargetGraphHashing` to show the hashes of the targets. These hashes are eagerly calculated when `TargetNode`s are created in parser and the are not used in other places.

This change removes `rawInputsHashCode` field from `TargetNode` and changes `TargetGraphHashing` to calculate hashes by getting raw attributes of `TargetNode`s from the parser and performing hashing of those attributes.

This also solves the problem with `TargetNode`s that have configured attributes since the factory used to use `hashCode` of the source `RawTargetNode`.

This logic uses async operations to avoid issues with starvation of the executor since logic was changed to compute attributes explicitly (which happens on the same executor that is used to compute target hashes).

Reviewed By: sbalabanov

fbshipit-source-id: 0a9416f2dd