Extract ParseStatsKeeper
Summary:
This extracts the tracking of parse events to a separate stats tracker
object. The console eventbus listeners get the current state of parsing
from this stats tracker. This better separates concerns.

One particularly obvious benefit of this is the ease of testing. For
reference, D13113481, is a change that fixed
SuperConsoleEventBusListener's handling of the multiple parse case.
Compare it's testing vs this changes
ParseStatsTrackerTest.testMultiParse().

Reviewed By: bobyangyf

fbshipit-source-id: 4679a0df6c