Add support for Buck LiveTemplates
Summary:
Add support for Buck LiveTemplates (which have support for
some language context), and move some of the identifier
autocompletion into those templates.

This can be used to help steer users to the syntax and forms
of statements/expressions that may be most useful to them.

For example, the `glob` template in this diff explicitly
fills in the `exclude` keyword (to steer users away from
using `excludes`, which we would like to deprecate/remove)
and `include_dotfiles=False` (to reduce confusion for users
who may not know what the default behavior of glob is).

Reviewed By: asp2insp

fbshipit-source-id: ea1bc6d612