Fix a bug in kotlin_library() when tools.javac is specified.
Summary:
`KotlincToJarStepFactory` adds a `JavacStep` to its
list of steps even if no `.java` files were specified in the list of
`srcs` to a `kotlin_library()` rule to account for the case where
some annotation processor generates `.java` files that need to
be compiled to create the output JAR for the rule. In practice,
this is fairly uncommon, so the `JavacStep` does not end up doing
any work most of the time because it is not given any `.java` files
to compile.

In general, this is innocuous if the in-memory `javax.tools.JavaCompiler`
is used for compilation in the `JavacStep`, as it appears to silently ignore
an invocation of `getTask()` with an empty `Iterable` of `compilationUnits`.

By comparison, a standard command-line implementation of `javac`
errors out if no `.java` source files are specified. In our case, this
means that the `JavacStep` added by `KotlincToJarStepFactory` will
fail when something like the following is added to the Buck project's
`.buckconfig` and there are no `.java` files passed to a `kotlin_library()`
rule (or generated via an annotation processor):

```
[tools]
  javac = /usr/bin/javac
```

This diff fixes `ExternalJavac` so that it behaves the same as
`javax.tools.JavaCompiler`, doing no compilation work when no `.java`
files are specified.

Reviewed By: jkeljo

fbshipit-source-id: f89b9e6c59