Replace BuildTargetPatternParser with a target base name in query logic
Summary:
`BuildTargetPatternParser` is used to pass data (target base name and whether wildcards are allowed), but not to parse patterns. This makes unnecessary dependency on this class in places not related to parsing.

This change replaces `BuildTargetPatternParser` with a target base name in multiple places.

A few notes that help explain this refactoring:

- `BuildTargetPatternParser` that is used in changed places is always created without allowing wildcards
- `BuildTargetPatternParser.fullyQualified` is replaced with an empty base name.

Reviewed By: jtorkkola

fbshipit-source-id: 18c5e2d468