Generate html and hoogle in single haddock command invocation
Summary:
We used to build html and hoogle with separate haddock command
invocations. Both builds from scratch and as a result it doubles the
step time.

In this change, we builds them in a single invocation, as a result,
the generated files will be in the same directory. If this is causes
problem, we can change haddock to put generated files in separate dirs.
But all of our existing use cases doesn't care about this or even expect
them to be in the same directory.

Reviewed By: xich

fbshipit-source-id: e13296eef6