Reduce error level of string parsing
Summary:
In `load()` statements, ideabuck can definitively parse paths to
files and determine whether or not a target is valid.  In other
contexts where a string is present, ideabuck is not yet smart
enough to know whether or not the string is intended to refer
to a target, so flagging it as an error is overly severe.  Change
the annotation level from "error" to "warning" accordingly.

This is part of an effort to make ideabuck's parsing more correct,
so that ideabuck errors are truly errors that must be fixed before
the user can build with buck.

Reviewed By: cute-jumper

fbshipit-source-id: 9a6d281ad7