Add reference grammar for next-gen ideabuck
Summary:
Add the end-goal for ideabuck's grammar, based on Python's
grammar with tweaks for the Starlark syntax.

Based on this grammar, ideabuck will undergo a series of
refactorings to move its grammar to this one.

Note that this file is not actually used in the build,
and there is no intention of doing so directly.  It serves
merely as a prototype to verify that the end goal makes
sense.

Reviewed By: asp2insp

fbshipit-source-id: ee204a07a3