Avoid using Config.GLOBAL in Ini
Summary: There are multiple places where `Ini` is created. By default, it would use a static `Config.GLOBAL` instance to configure itself. At the same time Buck uses custom configuration created in `Inis.makeIniParser` which is different from the global one (specifically, escaping is disabled so that a colon (in target name) doesn't need to have a backslash). Some places in tests don't use this custom configuration and, as a result, these tests may fail. What's more frustrating is that the test execution depends on the order of execution. Some tests may run through a path that would configuration the global instance and then the tests that don't use the custom configuration but use the global config work fine. When the same tests run in isolation (without any other tests before them), they would fail because the configuration is wrong. That would also mean that CI doesn't catch these failures.

Reviewed By: jtorkkola

fbshipit-source-id: 9ee095387a