Reapply: Don't ignore main cell's buck-out when creating ProjectFilesystem
Summary:
Reapply this diff which got reverted due to a "Buck not picking up changes" bug. This bug is fixed by an earlier diff in the stack, as is evidenced by the fact that `IntraCellIntegrationTest::testBuckdPicksUpChangesInChildCell` passes.

Original Commit Message:

This fixes a bug where Buck can't see it's generated artifacts when a child cell is contained within a parent cell and the parent references something in `child/buck-out`.

Reviewed By: cjhopman

fbshipit-source-id: abfa2624c5