Add support for select in Python build file parser
Summary:
This change adds support for `select` in Python build file parser.

This is implemented by creating `SelectorList` and `SelectorValue` objects in `select` method provided by the parser. These objects are then serialized in JSON as maps with special keyword `type` indicating the type of the object and a few additional entries to pass the actual data. These maps are then deserialized into `SelectorValue` and `ListWithSelects` in Java.

Reviewed By: bobyangyf

fbshipit-source-id: a180198ee3