Fix in-process Kotlin compilation on Java 9+
Summary:
In Java 9+, compiler classes are loaded using the platform class loader rather than a separate system tool class loader.

When using an in-process Kotlin compiler JAR, we create a child class loader of the system tool class loader, but this is null (which represents the bootstrap class loader) on Java 9+, so the Kotlin compiler can't access the Java compiler. Fix this. Also make sure other places in the codebase use the same codepath. Luckily, newer versions of the Kotlin compiler JARs already run on Java 11.

Reviewed By: styurin

fbshipit-source-id: ec197a375b