Fix bug where android_instrumentation_test generates incorrect paths
Summary:
This only happens in an intra-cell environment. If you have a `child-repo` that trys to `buck test parent//some/android:instrumentation-test`, then when Buck invokes java with it's packaged resources, it will give resource paths relative to `parent` rather than relative to `child-repo`. This causes errors when you run the tests.

This diff fixes the issue by providing an absolute path, as we do for the `testrunner_classes`. This ensures it works no matter where the test is invoked from.

Reviewed By: bobyangyf

fbshipit-source-id: c8cadd4177