Strip unnecessary archs from copied Swift libraries
Summary:
Starting sometime around Xcode 10.1 `swift-stdlib-tool` copies all the archs in the swift libraries, including those that are unused by the binary.  This is both inefficient size-wise, and also can cause App Store rejections.

This refactors `SwiftStdlibStep` to invoke `lipo` to strip out unused archs in the copied libraries by determining which archs the binary in the bundle uses.

Reviewed By: skotchvail

fbshipit-source-id: 81f8fee8ca