Do not upload 0 size manifests
Summary:
Usually this seems to imply something terrible has happened on the host
(see cache corruption issues with trying to gunzip 0-size files because
of cache corruption).

I think this is safe to do because the outputStream contains the
serialized Manifest object. The object is not null so we should be able
to guarantee that this file *must* be >0 bytes in size.

Throwing here so that we get logging, and it's okay to fail in this case
beacuse the alternative is uploading a corrupted artifact into the
cache.

Reviewed By: sbalabanov

fbshipit-source-id: 7a5dc045b2