Do not assume selector list contains a single element
Summary:
Python dsl parser incorrectly assumed that the selector list contains a single element.
In reality, it may contain combination of select() and other objects that can be concatenated to the result of the select().

Calling Iterables.getOnlyElement() when the list contains more than one element triggers a runtime exception.
Use Iterables.getLast() instead to determine the type of the selectable.

Reviewed By: styurin

fbshipit-source-id: 1a02bcf861