Adjust IntelliJ project to allow access to grammar artifacts
Summary:
The IntelliJ compiler preferences for the Buck project explicitly
disallow compilation of source files in buck-out/gen.  However,
for the ideabuck plugin, we *want* to compile the generated grammar
in `buck-out/gen/tools/psi-grammar` so that debugging the ideabuck
plugin in IntelliJ uses the same source as the plugin that buck
eventually builds.  Since `excludeFromCompile` does not appear to
allow whitelisting, make `buck-out/gen/tools/psi-grammar` available
by expanding the directories under `buck-out/gen/*` and excluding
the directories not used by ideabuck.

Reviewed By: IanChilds

fbshipit-source-id: e8ffd50c02