Exopackage for External Test Runner
Summary:
Currently exopackage installation does not work with the external test runner for AndroidInstrumentationTests.
This diff enables exopackage install in a nice minimal way by creating a symlink tree on disk for exactly how the exopackage directory should be laid out. This means that instead of taking a dependency on the exopackage machinery, the testrunner can simply mirror the files from that dir on to the device.

Reviewed By: styurin

fbshipit-source-id: 81f58a6e3b