Incorporate go.default_platform in platform resolution
Summary:
Currently, the way a Go platform is resolved considers the following in order:
1. Whether the build rule has a flavor/platform defined (e.g. //my/cool:rule#some-platform)
2. The platform defined by the rule (e.g. `go_binary`). Oftentimes this is defaulted to whatever platform is the default for the whole project.
3. The default "global" platform - again, this ends up being the default platform for the project.

This change adds a step between (1) and (2) that checks the value of `go.default_platform`. As far as I can tell, this is undocumented, but can be overriden via buck configs (`--config go.default_platform=foo-bar`) to change which platform Go uses. Given a definition like:
```
[go#my-cool-platform]
  root = /path/to/root
  os = darwin
  arch = amd64
```
you could, for example easily use `--config go.default_platform=my-cool-platform` to do a cross-compilation build for darwin.

Reviewed By: mkaczanowski

fbshipit-source-id: c98a58d0ea