Fix handling of output size and hash
Summary:
The OUTPUT_SIZE value should always be present. This is already
validated in some other places, ensure it here also.

The output hash should almost always be present (only not when exceeding
the hash size limit). We should report it whenever it is present. We are
not currently enforcing anywhere that it is present when we think it
should be, so I kept this as only a warning rather than a crash. This
would be a great place to report a soft error, but we have no support
for that.

Reviewed By: sbalabanov

fbshipit-source-id: 4295f227f3