Ignore version_control.generate_statistics config when deciding to restart daemon
Summary:
Buck defaults to restarting the daemon when config options change. However, there's a whitelist of config fields in `getIgnoreFieldsForDaemonRestart()` that do not cause the daemon to restart if they happen to change.

This diff adds the `version_control.generate_statistics` config option to this list. (This config option determines whether or not buck calls into version control during the build in order to log information about local file changes.) This option shouldn't have any affect on things like rule key computation, so I don't believe it should cause the daemon to restart.

## Context
See D13530300 for more info, but the TLDR is that the GraphQL Build Infra team wishes to disable `version_control.generate_statistics` for the buck invocations that run during `arc rebuild-graphql` since the vcs logging cause the build to hang frequently. But first, we want to land this diff to make sure we're not causing the daemon to restart unnecessarily.

Reviewed By: cjhopman

fbshipit-source-id: cf33148a52