Remove unnecessary field
Summary:
This is odd, this field is holding a type and then using the exact logic
for that type that was used when it's created to set or not set an
optional (and it's using that logic for interacting with the Optional).

Just look at if the Optional is set or not.

Reviewed By: jtorkkola

fbshipit-source-id: 8ab594b333