Extract NetworkStatsTracker
Summary:
This extracts the tracking of network events to a separate stats tracker
object. The console eventbus listeners get the current state of
uploads/downloads/cacherate from this stats tracker. This better
separates concerns.

One particularly obvious benefit of this is the ease of testing and not
tying testing of the tracking of these events with how we render them
(and in fact, currently we have to parse out the entire super-rendered
frame to get the network related stuff).

Reviewed By: bobyangyf

fbshipit-source-id: 5c11443c1a