Introduce UnresolvedCxxPlatform
Summary:
The CxxPlatform (and other logical toolchains) are required at 2
distinct parts of the build. During the parse/configuration phase the
CxxPlatform is required for us to determine parse time deps (i.e. for
Descriptions to implement findDepsFromConstructorArg, or, ideally in
the future, for the framework to do that derivation). During action
graph creation and execution the values of the CxxPlatform are needed to
control behavior and to reference tools and use them correctly.

These toolchains tend to consist of two things, Tools and static
configuration. As we want to support having the Tools be implemented by
binaries defined in build files, we've run into an issue, a Tool cannot
be resolved until after the parse/configuration phase. The way we've
dealt with that is that we have slowly hidden the Tools behind a
ToolProvider that exposes parseTimeDeps() and can then, with a
BuildRuleResolver be resolved to the real Tool during action graph
creation. For CxxPlatform specifically, we've converted almost all of
the Tools (but not quite all of them) in this way. We haven't come up
with a way for to shift the non-Tool configuration in this way (which
would be a requirement for moving the entire toolchain specification
into the build files), but we could maybe do something with Arg which
solves a similar problem.

This change takes the approach of hiding the entire CxxPlatform behind
an UnresolvedCxxPlatform that exposes getParseTimeDeps() and can be
resolved to the real CxxPlatform at action graph creation time.

This is actually a really simple conversion, except for the places where
the CxxPlatform is used in another (logical) toolchain. In those cases,
because the toolchains are created early, we need to either update
them or to resolve the CxxPlatform early. For now, I resort to
resolving the platform early. Due to the simplicitly of hiding the
entire toolchain behind a provider in this way, migrating those bad
uses is actually straightforward (and I do some of them in some
followups).

In a dependent diff, I use this to introduce a cxx_platform rule and the
change to .buckconfig so that you can move a full platform specification
into build files.

Reviewed By: bobyangyf

fbshipit-source-id: 04da51672c