Force Go testmain platform GOOS/GOARCH to local machine value
Summary: This diff forces the GOOS/GOARCH values when generating the build rule for the Go testmain binary. This is necessary because of the way Go tests are created - they use a testmain program (itself written in go) to generate the actual Go tests. In order to build tests in a cross-platform way (e.g. with `buck build -c go.default_platfomr=darwin_amd64 //some:target` on a linux system), for the tests to be generated and built (not run), testmain needs to be a binary suitable for the local machine.

Reviewed By: styurin, mkaczanowski

fbshipit-source-id: e485ea986f