treat testmaingen as part of goplatform
Summary:
testmaingen is a middlestep executable that generates go code during buck test. It should be treated as any other tool in go toolchain / goplatform.

This diff removes config test_main_gen (which is not platform specific) and moves the testmaingen to the goplatform. If the tool is not provided, it's going to be built from source (as it was so far)

Reviewed By: styurin

fbshipit-source-id: a8c5ff27a9