Add targets to App Module Metadata step
Summary: add the mapping of modules to the target they include into the metadata output.  Example outputs are updated in the unit tests.  Modules are constructed by including the packagable contents of the targets for each module.  The targets for the modules are determined by the seed targets and the exclusive dependencies of the seed targets. This is used to audit the targets used to construct modules in order to assist investigations into how modules are constructed and integrate into toolchains that analyze the module graph.

Reviewed By: styurin

fbshipit-source-id: 6ccbec8390