Remove entries from bazel jar instead of using jarjar
Summary: `bazel_deploy.jar` contains some of the classes that are already present in other libraries and which should be removed by jarjar, but my testing showed that jarjar doesn't remove those classes. Since jarjar is used to remove classes only, it can be replaced with `zip` commands that remove jar entries.

Reviewed By: bobyangyf

fbshipit-source-id: 7cbd9cf6d8