Add support for app extension schemes
Summary:
Apple uses `wasCreatedForAppExtension` in this position in the scheme XML to automatically begin debugging the bundled extension(s) when they're launch as part of the app debugging session.
```
<?xml version="1.0" encoding="UTF-8"?>
<Scheme
   LastUpgradeVersion = "1010"
   wasCreatedForAppExtension = "YES"
   version = "1.3">
```
Add implicit schemes for all app extensions bundled in the workspace source target. The scheme names will be of the form `AppSchemeName+ExtensionName`. This allows Xcode to build + run the new scheme to debug the app and the extension simultaneously without any intervention.

Reviewed By: williamtwilson

fbshipit-source-id: baa3d492e3