Introduce ParsingContext to reduce the amount of arguments passed to parsing methods
Summary:
Parser and surrounding infrastructure uses multiple parameters that often are passed together and maintaining these parameters is time consuming. For example, adding a new parameter or passing an existing parameter to a place it wasn't used requires updating multiple places.

The solution here is to introduce `ParsingContext` that contains parameters that may affect parsing in some way or used by the parsing infrastructure.

This class uses a builder pattern for creation of objects with some required parameters (cell and executor). The builder pattern helps with the problem of using primitive types for some parameters (booleans in this case), which be easily mixed up when passed as arguments in a method.

The downside of this approach is that some of the parameters may not be used by some of the methods and it might be difficult to tell is a given method is using a parameter, but IMO the benefits outweigh this problem.

More parameters will be added later to this class. To make it stable, build targets and target nodes are not included in it, they are still passed as arguments.

Reviewed By: jtorkkola

fbshipit-source-id: 65609026cb