Universalize Daemon object as state storage
Summary:
This change makes it so that Daemon object is initialized even for NO_BUCKD scenario and serves as a source of truth for global caches no matter if in daemon mode or not. This simplifies code and removes branching.
In next diffs I'll rename this class to smth like BuckState or GlobalState.

Reviewed By: bobyangyf

fbshipit-source-id: 9efba3bca1