Validate constraints in target_compatible_with to match the target platform
Summary:
`target_compatible_with` was used to filter out targets when building target using a pattern (for example, `//...`). Buck would ignore targets whose constraints are not compatible with the target platform.

This change introduces an additional check that would verify that a target is compatible with the target platform when doing regular parsing.

Note that this check needs to be ignored when doing the build using patterns and ignoring incompatible targets. The targets in this case need to be created, but they will be filtered out at a later stage.

Reviewed By: bobyangyf

fbshipit-source-id: 250db7948a