Make PipelineNodeCache cache all computed values during a build
Summary:
Right now, PipelineNodeCache will always lookup values first via
cache.lookupComputedNode(). This is written expected that that call is
fast, but that's not true (it checks for a whole bunch of possible
invalidation on every call). On a miss in that lookup it then will check
it's own jobs cache (which is fast).

Change this to instead use the fast-lookup cache first and insert all
values into it. This greatly speeds up parser performance in two
important scenarios: no-op builds for large graphs goes from 3-7s down
to <1.5s and creation of very, very large graphs decreased by as much as
30%.

Note that these are created as part of the PerBuildState (not the
Parser) and so are thrown away after the parse part of a command is
finished.

Reviewed By: bobyangyf

fbshipit-source-id: 9a78925db0