Fix parallelization in target graph hashing
Summary:
When "finishing" nodes during target graph hashing, we were using a
direct executor when transforming the deps results. This is the final
step fo each of these futures, it's happening as we traverse up the
graph and so necessarily causes our originally parallel computation to
converge to using a single thread.

Instead, change this to transform on our executor to continue using the
configured parallelism.

"finishing" these nodes takes somewhere between 30-50% of target hash
time and this significantly speeds that up for large graphs.

Reviewed By: bobyangyf

fbshipit-source-id: e32cfb3c22