Add a fast path to ignore collection types for AttrFilterFunction
Summary:
This invokes filterAttributeContents() which will traverse the entire
type hierarchy of each attr (see the various type coercers). Collection
types are (1) very common (2) expensive to convert to string and (3) we
shouldn't apply the filter to the stringified form, and so we have a
fast path to ignore them.

In a large graph's action graph creation, converting collections to
strings accounted for something like 40% of allocations.

Reviewed By: styurin

fbshipit-source-id: 9a07271c31