Optimize `buck project`'s LazyActionGraph
Summary:
In getActionGraphBuilderWhileRequiringSubgraph(), to traverse a part of
the TargetGraph, the LazyActionGraph is currently calling getSubGraph().
That call does a significant amount of work.

In addition, it is then traversing that entire subgraph to call
requireRule on each node. When this function is then called many times
where there is some overlap in the subgraphs, it's wasted work (and
algorithmically quite poor) to traverse nodes that we know have already
been required.

This optimizes that function to (1) just traverse the subgraph in the
original TargetGraph and (2) short-circuit traversal for things we've
already traversed.

Reviewed By: sbalabanov

fbshipit-source-id: 4a6517dd38