Add mechanism to configure a dynamic cxx_toolchain
Summary:
This adds the buckconfig support for specifying that a cxx platform is
configured by a cxx_toolchain target. The implementation side of the change is
pretty straightforward, the hard work there is done in previous diffs.

The tests are a bit more complicated because we require that the cxx_toolchain
define the entire toolchain. We could do the work to implement augmentation,
but it'd be ugly and complicated. Instead, if we wait for ProvidesCxxPlatform
to be changed to CxxToolchainProvider and for extensions to exist we can just
convert internal stuff to be exposed as CxxToolchainProvider and we get really
clean augmentation of whatever we want (along with the other benefits of those
framework-y changes).

Reviewed By: sbalabanov

fbshipit-source-id: 39d674964c